+++ CHANGELOG.md
@@ -11,7 +11,9 @@
 - CLI: new `mc mobile-app-instances` command [#252]
 - CLI: new `mc mobile-apps` command [#252]
 - CLI: new `mc notifications` command for MindSphere notifications [#252]
+- CLI: new `mc signal-calculation` command for MindSphere Signal Calculation Service [#258]
 - CLI: fixed bug in `mc aggregates` command with multiple lines of same aggregate
+- SDK: signal calculation client [#258]
 - SDK: Job Manager Client [#255]
 - SDK: Data Exchange Client [#256]
 - SDK: NotficationV4 Client [#252]